+++ edsnlp/pipelines/ner/scores/emergency/ccmu/factory.py
@@ -19,9 +19,16 @@ DEFAULT_CONFIG = dict(
 
 
 @deprecated_factory(
-    "emergency.ccmu", "eds.emergency.ccmu", default_config=DEFAULT_CONFIG
+    "emergency.ccmu",
+    "eds.emergency.ccmu",
+    default_config=DEFAULT_CONFIG,
+    assigns=["doc.ents", "doc.spans"],
 )
-@Language.factory("eds.emergency.ccmu", default_config=DEFAULT_CONFIG)
+@Language.factory(
+    "eds.emergency.ccmu",
+    default_config=DEFAULT_CONFIG,
+    assigns=["doc.ents", "doc.spans"],
+)
 def create_component(
     nlp: Language,
     name: str,