+++ edsnlp/pipelines/ner/scores/sofa/factory.py
@@ -17,8 +17,17 @@ DEFAULT_CONFIG = dict(
 )
 
 
-@deprecated_factory("SOFA", "eds.SOFA", default_config=DEFAULT_CONFIG)
-@Language.factory("eds.SOFA", default_config=DEFAULT_CONFIG)
+@deprecated_factory(
+    "SOFA",
+    "eds.SOFA",
+    default_config=DEFAULT_CONFIG,
+    assigns=["doc.ents", "doc.spans"],
+)
+@Language.factory(
+    "eds.SOFA",
+    default_config=DEFAULT_CONFIG,
+    assigns=["doc.ents", "doc.spans"],
+)
 def create_component(
     nlp: Language,
     name: str,