+++ edsnlp/pipelines/qualifiers/family/factory.py
@@ -15,8 +15,17 @@ DEFAULT_CONFIG = dict(
 )
 
 
-@deprecated_factory("family", "eds.family", default_config=DEFAULT_CONFIG)
-@Language.factory("eds.family", default_config=DEFAULT_CONFIG)
+@deprecated_factory(
+    "family",
+    "eds.family",
+    default_config=DEFAULT_CONFIG,
+    assigns=["span._.family"],
+)
+@Language.factory(
+    "eds.family",
+    default_config=DEFAULT_CONFIG,
+    assigns=["span._.family"],
+)
 def create_component(
     nlp: Language,
     name: str,