+++ edsnlp/pipelines/qualifiers/negation/factory.py
@@ -18,8 +18,17 @@ DEFAULT_CONFIG = dict(
 )
 
 
-@deprecated_factory("negation", "eds.negation", default_config=DEFAULT_CONFIG)
-@Language.factory("eds.negation", default_config=DEFAULT_CONFIG)
+@deprecated_factory(
+    "negation",
+    "eds.negation",
+    default_config=DEFAULT_CONFIG,
+    assigns=["span._.negation"],
+)
+@Language.factory(
+    "eds.negation",
+    default_config=DEFAULT_CONFIG,
+    assigns=["span._.negation"],
+)
 def create_component(
     nlp: Language,
     name: str,